slm002/deprecated/deprecated_templates/requirements.txt
pillow>=10.0.0
requests>=2.31.0
psutil>=5.9.0
numpy>=1.24.0